# pkg/trace: reduce allocs made by the obfuscator