Reduce number of allocs by using the tokenizer internal buffer for filtering operations.

Benchmarks:
```
name                               old time/op    new time/op    delta
ObfuscateSQLString/Escaping/512-8    8.32µs ± 2%    7.80µs ± 5%   -6.35%  (p=0.000 n=18+5)
ObfuscateSQLString/Grouping/199-8    4.24µs ± 1%    3.85µs ± 1%   -9.18%  (p=0.000 n=18+5)
ObfuscateSQLString/Large/3694-8      71.8µs ± 2%    67.5µs ± 2%   -5.93%  (p=0.000 n=16+5)
ObfuscateSQLString/Complex/969-8     19.4µs ± 4%    18.1µs ± 1%   -6.50%  (p=0.000 n=20+5)
ObfuscateSQLString/random-8          1.27µs ± 3%    1.21µs ± 1%   -4.70%  (p=0.000 n=20+5)

name                               old alloc/op   new alloc/op   delta
ObfuscateSQLString/Escaping/512-8    1.95kB ± 0%    1.25kB ± 0%  -35.90%  (p=0.000 n=20+5)
ObfuscateSQLString/Grouping/199-8      828B ± 0%      560B ± 0%  -32.37%  (p=0.000 n=20+5)
ObfuscateSQLString/Large/3694-8      16.3kB ± 0%    11.3kB ± 0%  -30.60%  (p=0.000 n=20+5)
ObfuscateSQLString/Complex/969-8     3.01kB ± 0%    3.01kB ± 0%   -0.17%  (p=0.000 n=20+5)
ObfuscateSQLString/random-8            256B ± 0%      248B ± 0%   -3.12%  (p=0.000 n=20+5)

name                               old allocs/op  new allocs/op  delta
ObfuscateSQLString/Escaping/512-8      23.0 ± 0%       6.0 ± 0%  -73.91%  (p=0.000 n=20+5)
ObfuscateSQLString/Grouping/199-8      22.0 ± 0%       6.0 ± 0%  -72.73%  (p=0.000 n=20+5)
ObfuscateSQLString/Large/3694-8         150 ± 0%         6 ± 0%  -96.00%  (p=0.000 n=20+5)
ObfuscateSQLString/Complex/969-8       11.0 ± 0%       6.0 ± 0%  -45.45%  (p=0.000 n=20+5)
ObfuscateSQLString/random-8            9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=20+5)
```